frontend/dx/backends/django/requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
distlib==0.3.3
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
django-heroku==0.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
filelock==3.4.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
platformdirs==2.4.0
promise==2.3
PyJWT==2.3.0
pytz==2021.3
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
whitenoise==5.3.0
